requirements.txt
torch>=2.1.0
timm>=0.9.0
fastapi>=0.108.0
uvicorn>=0.25.0
python-multipart>=0.0.6
Pillow>=10.0.0
numpy>=1.24.0
